add navbar design
