+++ src/Componenets/GameList.jsx
@@ -1,26 +1,11 @@
-import { useSelector } from "react-redux";
-import { useState, useEffect } from 'react';
 
-export default function GameList() {
-  const done = useSelector((state) => state.data.done);
-  const list = useSelector((state) => state.data.items);
-  const [filteredArr, setfilter] = useState([]);
 
-  useEffect(() => {
-    setfilter([...list]);
-  }, [list]);
-
-  const searchHandler = (e) => {
-    const search = e.target.value;
-    const copy = list.filter(item => item.title.includes(search));
-    setfilter(copy);
-  };
+export default function GameList({Done, Games}) {
 
   return (
-    <div>
-      <input type="text" onChange={searchHandler} />
-      {done ? (
-        filteredArr.map((item) => {
+    <div >
+      {Done ? (
+        Games.map((item) => {
           return <div key={item.id}> {item.title}</div>;
         })
       ) : (
