+++ src/Layout/Navbar.module.css
@@ -1,0 +1,44 @@
+@import url('https://fonts.googleapis.com/css2?family=Fredoka+One&display=swap');
+
+.upbar {
+  width: 100%;
+  background-color: #262626;
+  color: white;
+  height: 70px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 15 15px;
+  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.2);  
+}
+
+.logo {
+  font-family: 'Fredoka One', cursive;
+  font-size: 1.7rem;
+  font-weight: 700;
+  margin: 0;
+  margin-left: 16px;
+}
+
+.searchbar {
+  width: 50%;
+  height: 40px;
+  border-radius: 7px;
+  border: none;
+  padding: 0 20px;
+  font-size: 1rem;
+}
+
+.searchbaricon {
+  color: white;
+  font-size: 35px;
+  margin-left: 10px;
+  margin-right: 10px;
+}
+
+.right {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+}
